# Application_Prototype/mvp_v1/backend/openapi.json
{"openapi":"3.1.0","info":{"title":"AI-Compass API","version":"1.0.0"},"paths":{"/api/v1/questionnaire/":{"get":{"tags":["Questionnaire"],"summary":"Get Questionnaire","description":"Fetch the full questionnaire with all questions and answers.","operationId":"get_questionnaire_api_v1_questionnaire__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Questionnaire"}}}}}}},"/api/v1/companies/":{"post":{"tags":["Companies"],"summary":"Create Company","description":"Create a new company profile.","operationId":"create_company_api_v1_companies__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/responses/":{"post":{"tags":["Responses"],"summary":"Create Response","description":"Initialize a new assessment response for a company.","operationId":"create_response_api_v1_responses__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/responses/{response_id}/items":{"patch":{"tags":["Responses"],"summary":"Save Answer","description":"Save or update an answer (Autosave).","operationId":"save_answer_api_v1_responses__response_id__items_patch","parameters":[{"name":"response_id","in":"path","required":true,"schema":{"type":"integer","title":"Response Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/responses/{response_id}/complete":{"post":{"tags":["Responses"],"summary":"Complete Assessment","description":"Mark assessment as complete, update company details, persist data to DB, and trigger scoring/analysis.","operationId":"complete_assessment_api_v1_responses__response_id__complete_post","parameters":[{"name":"response_id","in":"path","required":true,"schema":{"type":"integer","title":"Response Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseComplete"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/responses/{response_id}/results":{"get":{"tags":["Results"],"summary":"Get Results","description":"Retrieve full analysis results.\n1. Fetch Session Data\n2. Score Dimensions (Replicating DataPipeline logic)\n3. Run ML Inference\n4. Return JSON","operationId":"get_results_api_v1_responses__response_id__results_get","parameters":[{"name":"response_id","in":"path","required":true,"schema":{"type":"integer","title":"Response Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/responses/{response_id}/pdf":{"get":{"tags":["Results"],"summary":"Generate Pdf","description":"Generate and download PDF report.","operationId":"generate_pdf_api_v1_responses__response_id__pdf_get","parameters":[{"name":"response_id","in":"path","required":true,"schema":{"type":"integer","title":"Response Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AnswerSchema":{"properties":{"answer_id":{"type":"integer","title":"Answer Id"},"answer_text":{"type":"string","title":"Answer Text"},"answer_level":{"type":"integer","title":"Answer Level"},"answer_weight":{"type":"number","title":"Answer Weight"}},"type":"object","required":["answer_id","answer_text","answer_level","answer_weight"],"title":"AnswerSchema"},"CompanyCreate":{"properties":{"company_name":{"type":"string","title":"Company Name"},"industry":{"type":"string","title":"Industry"},"website":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Website"},"number_of_employees":{"type":"string","title":"Number Of Employees"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"}},"type":"object","required":["company_name","industry","number_of_employees"],"title":"CompanyCreate"},"CompanyResponse":{"properties":{"company_name":{"type":"string","title":"Company Name"},"industry":{"type":"string","title":"Industry"},"website":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Website"},"number_of_employees":{"type":"string","title":"Number Of Employees"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"company_id":{"type":"integer","title":"Company Id"}},"type":"object","required":["company_name","industry","number_of_employees","company_id"],"title":"CompanyResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"QuestionSchema":{"properties":{"question_id":{"type":"integer","title":"Question Id"},"dimension_id":{"type":"integer","title":"Dimension Id"},"dimension_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dimension Name"},"header":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Header"},"question_text":{"type":"string","title":"Question Text"},"type":{"type":"string","title":"Type"},"weight":{"type":"number","title":"Weight"},"optional":{"type":"boolean","title":"Optional"},"answers":{"items":{"$ref":"#/components/schemas/AnswerSchema"},"type":"array","title":"Answers","default":[]}},"type":"object","required":["question_id","dimension_id","question_text","type","weight","optional"],"title":"QuestionSchema"},"Questionnaire":{"properties":{"questions":{"items":{"$ref":"#/components/schemas/QuestionSchema"},"type":"array","title":"Questions"}},"type":"object","required":["questions"],"title":"Questionnaire"},"ResponseComplete":{"properties":{"company_details":{"$ref":"#/components/schemas/CompanyCreate"}},"type":"object","required":["company_details"],"title":"ResponseComplete"},"ResponseCreate":{"properties":{"company_id":{"type":"integer","title":"Company Id"}},"type":"object","required":["company_id"],"title":"ResponseCreate"},"ResponseDetail":{"properties":{"response_id":{"type":"integer","title":"Response Id"},"company_id":{"type":"integer","title":"Company Id"},"total_score":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Total Score"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"cluster_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Cluster Id"}},"type":"object","required":["response_id","company_id","created_at"],"title":"ResponseDetail"},"ResponseUpdate":{"properties":{"question_id":{"type":"integer","title":"Question Id"},"answer_ids":{"items":{"type":"integer"},"type":"array","title":"Answer Ids"}},"type":"object","required":["question_id","answer_ids"],"title":"ResponseUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}